Separate convert function to platform specific files
